Add support for 24/7 hours
The protocol now provides a `twentyfourseven` flag to cover the most
common scenario as an alternative to `regularHours` (they are mutually
exclusive).
There seems to be a hole in the specification though, because
`twentyfourseven` could potentially be defined but `false`: in that
scenario the conversion will fail, throwing an exception.
